+++ packages/xchain-bitcoincash/src/utils.ts
@@ -4,7 +4,18 @@ import * as bchaddr from 'bchaddrjs'
 import coininfo from 'coininfo'
 import { Address, Balance, Fees, Network, Tx, TxFrom, TxParams, TxTo } from '@xchainjs/xchain-client'
 import { AssetBCH, BaseAmount, baseAmount } from '@xchainjs/xchain-util/lib'
-import { DerivePath, FeeRate, FeeRates, FeesWithRates, Transaction, AddressParams, UTXOs, UTXO } from './types'
+import {
+  DerivePath,
+  FeeRate,
+  FeeRates,
+  FeesWithRates,
+  Transaction,
+  AddressParams,
+  UTXOs,
+  UTXO,
+  TransactionInput,
+  TransactionOutput,
+} from './types'
 import { getAccount, getRawTransaction, getUnspentTransactions } from './haskoin-api'
 import { Network as BCHNetwork, TransactionBuilder } from './types/bitcoincashjs-types'
 
@@ -114,14 +125,19 @@ export const bchNetwork = (network: Network): BCHNetwork => {
 }
 
 /**
- * Get address prefix based on the network.
- * Convert to Legacy Address.
- *
- * @param {string} network
- * @returns {string} The address prefix based on the network.
- *
+ * BCH new addresses strategy has no any prefixes.
+ * Any possible prefixes at the TX addresses will be stripped out with parseTransaction
  **/
-export const getPrefix = (network: string) => (network === 'testnet' ? 'm' : '1')
+export const getPrefix = () => ''
+
+/**
+ * Strips bchtest or bitcoincash prefix from address
+ *
+ * @param {Address} address
+ * @returns {Address} The address with prefix removed
+ *
+ */
+export const stripPrefix = (address: Address): Address => address.replace(/(bchtest:|bitcoincash:)/, '')
 
 /**
  * Convert to Legacy Address.
@@ -131,6 +147,16 @@ export const getPrefix = (network: string) => (network === 'testnet' ? 'm' : '1'
  */
 export const toLegacyAddress = (address: Address): Address => {
   return bchaddr.toLegacyAddress(address)
+}
+
+/**
+ * Convert to Cash Address.
+ *
+ * @param {Address} address
+ * @returns {Address} Cash address.
+ */
+export const toCashAddress = (address: Address): Address => {
+  return bchaddr.toCashAddress(address)
 }
 
 /**
@@ -144,20 +170,22 @@ export const parseTransaction = (tx: Transaction): Tx => {
   return {
     asset: AssetBCH,
     from: tx.inputs
-      .filter((input) => !!input.address)
+      // For correct type inference `Array.prototype.filter` needs manual type guard to be defined
+      .filter((input): input is Omit<TransactionInput, 'address'> & { address: string } => !!input.address)
       .map(
         (input) =>
           ({
-            from: input.address,
+            from: stripPrefix(input.address),
             amount: baseAmount(input.value, BCH_DECIMAL),
           } as TxFrom),
       ),
     to: tx.outputs
-      .filter((output) => !!output.address)
+      // For correct type inference `Array.prototype.filter` needs manual type guard to be defined
+      .filter((output): output is Omit<TransactionOutput, 'address'> & { address: string } => !!output.address)
       .map(
         (output) =>
           ({
-            to: output.address,
+            to: stripPrefix(output.address),
             amount: baseAmount(output.value, BCH_DECIMAL),
           } as TxTo),
       ),